Build out the weight Adjuster prediction
Remove the old named data sets and create 3 models. One is a statsmodel logestic regression which tells us how well their model fits. The next two are both scikit logistic regressions and their accuracy score at the end is based on A training set and a test set. We obtain their score in order to prove which model is the most accurate. We expect the balanced weight one to be the most accurate as itll auto balance itself with coefficients but if the second model is close we have converged to the best weights